refactor(react-grid): move expanded groups calculation to reducers (#229)
